#68 api 응답 모듈화 및, 프로필 카드, 구인글 카드 api 응답 경로 수정
