+++ src/types/api-res-error.d.ts
@@ -1,0 +1,6 @@
+/* --------------- 에러 --------------- */
+//에러가 나타났을 때 오는 코드
+interface ErrorResponse {
+  code: string;
+  message: string;
+}
