+++ src/types/api-res-upload.d.ts
@@ -1,0 +1,5 @@
+/* --------------- 파일업로드 --------------- */
+// POST /api/v1/uploads/images
+interface UploadedImage {
+  image_url: string;
+}
